+++ NEWS.md
@@ -1,3 +1,5 @@
+# testthat (development version)
+
 # testthat 3.1.10
 
 * Fix for upcoming R-devel release.