+++ NEWS.md
@@ -1,5 +1,8 @@
 # testthat (development version)
 
+* `expect_setequal()` correctly displays results when only one of actual and
+  expected is missing values (#1835).
+
 # testthat 3.1.10
 
 * Fix for upcoming R-devel release.