+++ NEWS.md
@@ -2,6 +2,9 @@
 
 * Helpers should no longer be run twice.
 
+* `test_file()` gains a `desc` argument which allows you to run a single 
+  test from a file (#1776).
+
 * `expect_setequal()` correctly displays results when only one of actual and
   expected is missing values (#1835).