+++ NEWS.md
@@ -1,5 +1,7 @@
 # testthat (development version)
 
+* Helpers should no longer be run twice.
+
 * `test_file()` gains a `desc` argument which allows you to run a single 
   test from a file (#1776).