+++ NEWS.md
@@ -1,5 +1,8 @@
 # testthat (development version)
 
+* `options(rlang_interactive = TRUE)` no longer causes `skip_on_cran()` to
+  not run on CRAN (#1868).
+
 * `skip_if_offline()` now errors if you don't have curl installed (#1854).
 
 * All packages, regardless of whether or not they use rlang 1.0.0, now