+++ NEWS.md
@@ -3,6 +3,17 @@
 * The skip column is now two characters wide to have a consistent width in case
   10 or more tests are skipped in a single file (@mgirlich, #1844).
 
+* `skip_if_offline()` now errors if you don't have curl installed (#1854).
+
+* All packages, regardless of whether or not they use rlang 1.0.0, now
+  use the new snapshot display for errors, warnings, and messages (#1856).
+
+* testthat uses an improved algorithm for finding the srcref associated with
+  an expectation/error/warning/skip. It now looks for the most recent call
+  that has known source and is found inside the `test_that()` call. This
+  generally gives more specific locations than the previous approach and
+  gives much better locations if an error occurs in an exit handler.
+
 * Helpers should no longer be run twice.
 
 * `test_file()` gains a `desc` argument which allows you to run a single 
@@ -48,9 +59,6 @@
 * `skip_on_cran()` no longer skips (errors) when run interactively.
 
 * `teardown_env()` works in more cases.
-
-* All packages, regardless of whether or not they use rlang, now
-  use the new snapshot display for errors, warnings, and messages.
 
 * testthat no longer truncates tracebacks and uses rlang's default tree
   display.