+++ NEWS.md
@@ -1,5 +1,8 @@
 # testthat (development version)
 
+* `options(rlang_interactive = TRUE)` no longer causes `skip_on_cran()` to
+  not run on CRAN (#1868).
+
 * The skip column is now two characters wide to have a consistent width in case
   10 or more tests are skipped in a single file (@mgirlich, #1844).