+++ NEWS.md
@@ -3,6 +3,10 @@
 * `expect_snapshot_value()` has an improved error if the object can't be
   safely serialized using the specified `style` (#1771).
 
+* `expect_no_error()`, `expect_no_warning()`, `expect_no_message()`, 
+  `expect_no_condition()`, `local_mocked_bindings()`, and 
+  `with_mocked_bindings()` are now stable, not experimental.
+
 * `StopReporter()` gains the ability to suppress praise when a test passes.
 
 * `expect_snapshot_file()` now generates clickable links to review changes